services/auth/requirements.txt
fastapi>=0.110.0
uvicorn>=0.30.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
passlib>=1.7.4
python-jose>=3.3.0
python-multipart>=0.0.6
pyotp>=2.8.0
python-dotenv>=1.0.0
email-validator>=2.0.0
psycopg2-binary>=2.9.9
